+++ packages/algoliasearch-helper/src/requestBuilder.js
@@ -277,9 +277,12 @@ var requestBuilder = {
       .sort()
       .forEach(function (facetName) {
         var facetValues = facetsRefinements[facetName] || [];
-        facetValues.sort().forEach(function (facetValue) {
-          facetFilters.push(facetName + ':' + facetValue);
-        });
+        facetValues
+          .slice()
+          .sort()
+          .forEach(function (facetValue) {
+            facetFilters.push(facetName + ':' + facetValue);
+          });
       });
 
     var facetsExcludes = state.facetsExcludes || {};
@@ -302,9 +305,12 @@ var requestBuilder = {
         }
         var orFilters = [];
 
-        facetValues.sort().forEach(function (facetValue) {
-          orFilters.push(facetName + ':' + facetValue);
-        });
+        facetValues
+          .slice()
+          .sort()
+          .forEach(function (facetValue) {
+            orFilters.push(facetName + ':' + facetValue);
+          });
 
         facetFilters.push(orFilters);
       });