+++ packages/react-instantsearch-core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.5.3](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.5.2...react-instantsearch-core@7.5.3) (2024-01-23)
+
+
+### Bug Fixes
+
+* **hydration:** generate cache with search parameters from server-side request ([#5991](https://github.com/algolia/instantsearch/issues/5991)) ([968cf43](https://github.com/algolia/instantsearch/commit/968cf433511eb172acee8eaf6a7c90740f7ee0a6))
+
+
+
+
+
 ## [7.5.2](https://github.com/algolia/instantsearch/compare/react-instantsearch-core@7.5.1...react-instantsearch-core@7.5.2) (2024-01-16)
 
 **Note:** Version bump only for package react-instantsearch-core